refactor: 사용치 않는 인터페이스 제거
